Change the python requirement syntax of test files
The purpose of this patch is to provide a way to create test files for more
specific Python versions, e.g. for testing async/await syntax in Python 3.5+.

In the new syntax we will evaluate the first leading comment line with
"python_version" in it.
python_version is equal to platform.python_version()[:3].