Fix ignore-names option initialization
cls.ignore_names starts life as a list, which is used as the default
value for the ignore-names option, and then it becomes a frozenset()
once the option is set.

Unfortunately, when flake8's --config option is used, it appears we run
those this sequence more than once, and that results in the frozenset()
being passed as the option's default value, which breaks flake8.

This change moves the default ignore-names list to a module-level
constant alongside the other default lists. This avoids the problem
described above and has the nice side-effect of making the code more
consistent.