Don't offset lineno when reporting errors for Python >= 3.8
In Python 3.8, bpo-33211 addressed what is considered a bug where a
function or a class with a decorator would report an incorrect `lineno`
and `col_offset`.  In versions < 3.8, the first decorator line was
reported, which requires the accounted for offset.  Now, the actual
`lineno` and `col_offset` is reported by the `ast` module.

Please see https://bugs.python.org/issue33211 for reference.