Treat "acronym" imports as their own error code
While not the recommended style, in practice, it is common for CamelCase
names from some packages to be imported as acronyms. Some examples:

    from BeautifulSoup import BeautifulSoup as BS
    from xml.etree.ElementTree import ElementTree as ET

We still consider these naming errors but now use a dedicated code
(N817) so that users can selectively enable or disable this rule
separate from N814 ("camelcase imported as constant").