Improve known metaclass base class detection
We previously only looked for known metaclass base classes that were
inheriting using the simple `ast.Name` syntax:

    class Meta(type): pass
    class Meta(ABCMeta): pass

We weren't accounting for the `ast.Attribute` syntax:

    class Meta(abc.ABCMeta): pass

This second syntax will now also match `ABCMeta` and be recognized as a
metaclass base class.

We could alternatively add `abc.ABCMeta` to METACLASS_BASES and
construct the fully-qualified name of each base class that uses the
`ast.Attribute` syntax. The simpler approach used here is currently
unambiguous, so we'll start with this.