Only run selected checks
We previously always ran all of our checks regardless of the configured
style guide (the set of selected and ignore error codes). Now, we build
our list of visitor classes based on which error codes they can emit and
which error codes are selected in the current style guide.

This is a (minor) performance optimization for folks who have many files
and only a subset of our checks enabled.