Pass 'parents' as a parameter when walking the tree
Rather than storing the 'parents' deque as an instance attribute,
pass it as a parameter as we walk the tree.

This gives us more explicit control of this stack's lifetime (and no
more mutable state on the object itself), and it allows us to better
restrict the typed interface we present to visitor methods (`Iterable`).

It also might result in a micro-optimization because read a parameter
value is less involved (faster) than loading an instance attribute, but
that isn't the goal of this change.