Refactor ignored name handling (#240)
We previously constructed a frozen set of names that should be ignored
at setup-time and then passed that set as an argument to each visitor
function. A helper function -- _ignored(name, set) -- was used to test
whether a given name should be ignored, matched using fnmatchcase().

This change introduces a `NameSet` type, which inherits from frozenset
and implements `__contains__` using fnmatchcase(). This allows us to
simply use the `in` operator in visitor functions.

The previous code also (sometimes) treated the `ignores` argument as
optional (defaulting to None). This value always exists so remove the
optional notation.